requirements.txt
flask
openai
python-dotenv
requests

# requirements.txt

flask==2.3.3
google-generativeai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
